added unit tests for generating chainid from different inputs
