+++ chain.go
@@ -6,6 +6,7 @@ package factom
 
 import (
 	"bytes"
+	"crypto/sha256"
 	"encoding/hex"
 	"encoding/json"
 	"errors"
@@ -48,6 +49,28 @@ func NewChainFromStrings(content string, extids ...string) *Chain {
 	e := NewEntryFromStrings("", content, extids...)
 	c := NewChain(e)
 	return c
+}
+
+// ChainIDFromFields computes a ChainID based on the binary External IDs of that
+// Chain's First Entry.
+func ChainIDFromFields(fields [][]byte) string {
+	hs := sha256.New()
+	for _, id := range fields {
+		h := sha256.Sum256(id)
+		hs.Write(h[:])
+	}
+	cid := hs.Sum(nil)
+	return hex.EncodeToString(cid)
+}
+
+// ChainIDFromStrings computes the ChainID of a Chain Created with External IDs
+// that would match the given string (in order).
+func ChainIDFromStrings(fields []string) string {
+	var bin [][]byte
+	for _, str := range fields {
+		bin = append(bin, []byte(str))
+	}
+	return ChainIDFromFields(bin)
 }
 
 // ChainExists returns true if a Chain with the given chainid exists within the
